Support the Community Highlander status hooks (#1)
Adds a check for whether the Community Highlander is active and meets
the minimum version of 1.19. If that's the case, this change makes the
soldier list items use the highlander's `GetPersonnelStatusSeparate()`
function for rendering soldiers' statuses.

This makes the mod compatible with any other mod that uses those hooks
to add or modify soldier statuses.

Closes #1.