Fix Days -> Hours for CHL integration (#1)
The Community Highlander doesn't automatically convert days to hours in
status strings, so this adds an event listener for the
`OverridePersonnelStatusTime` event to perform that conversion. This
ensures that the behaviour with and without the CHL is identical.

Note that this change means the mod must be compiled against the
Community Highlander.